refactor(geomap): move the Nominatim geocoder into its own module
geocoding.ts is left holding the contract and the registry; how one
provider talks to its service lives beside it rather than inside it, so a
second provider is a new module rather than more of this one.

The spec follows the implementation and asks for the provider directly
instead of resolving the default out of the registry, since what it tests
is Nominatim rather than whichever provider is default.
